sdk/python/foundation-models/stable-diffusion-3.5/sd35-image-to-image/requirements.txt
# Copyright 2024 Stability AI and The HuggingFace Team. All rights reserved.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

accelerate==1.1.1
diffusers==0.31.0
Flask==3.1.0
joblib==1.4.2 # Used by only score.py
pillow==11.0.0
protobuf==5.28.3
sentencepiece==0.2.0
transformers==4.46.2

# PyTorch

# NVIDIA GPUs: Uncomment the following 4 lines if you have a CUDA-capable GPU (CUDA version 11.8)
#torch
#torchvision
#torchaudio
#--extra-index-url https://download.pytorch.org/whl/cu118

# MacOS: Uncomment the following 4 lines (prerelease versions): https://developer.apple.com/metal/pytorch/
torch
torchvision
torchaudio
--extra-index-url https://download.pytorch.org/whl/nightly/cpu
